Implement the presentation layer and integrate functional repositories for the FarmPulse application. This commit transitions the UI from placeholders to a comprehensive Jetpack Compose design system and wires the view models to functional data services.
- Implement the UI design system with a custom typography scale and semantic color palette (Forest Green, Background Off-white, etc.) in `Theme.kt`.
- Create a feature-complete `HomeScreen` with a weather hero, real-time stat chips, horizontal hourly forecast, and an AI-powered insight card.
- Develop the `ForecastScreen` featuring a 7-day forecast breakdown and a rain probability bar chart.
- Build the `ScannerScreen` with a multipart image upload interface, usage quota tracking, and a detailed `AnalysisResultScreen` showing tree health distributions and AI recommendations.
- Establish the `HistoryScreen` with a persistent list of past analyses fetched from the local Room database.
- Transition `WeatherRepository` and `TreeRepository` from placeholders to functional implementations with offline-first caching and network status monitoring.
- Implement `WeatherViewModel`, `ForecastViewModel`, `ScannerViewModel`, and `HistoryViewModel` using Hilt for dependency injection and StateFlow for UI state management.
- Configure a centralized `MainScreen` with a persistent bottom navigation bar and customized slide-based transition animations between features.
- Integrate functional utility classes including `LocationHelper` for GMS location services and `NetworkMonitor` for real-time connectivity tracking.
- Add `farmpulse_ui_structure.md` and `farmpulse_screen_mockups.html` as comprehensive design and architectural references.
- Clean up the project by deleting the obsolete `MainActivity` placeholder in the legacy package.

+++ app/src/main/java/co/farmpulse/app/data/remote/dto/CurrentWeatherDto.kt
@@ -6,9 +6,8 @@ data class CurrentWeatherDto(
     @SerializedName("time") val time: String? = null,
     @SerializedName("temperature") val temperature: Double? = null,
     @SerializedName("wind_speed") val windSpeed: Double? = null,
-    @SerializedName("wind_direction") val windDirection: Int? = null,
+    @SerializedName("wind_direction") val windDirection: Double? = null,
     @SerializedName("condition_code") val conditionCode: String? = null,
     @SerializedName("icon") val icon: String? = null,
     @SerializedName("icon_path") val iconPath: String? = null
 )
-
