+++ app/src/main/java/co/farmpulse/app/data/remote/dto/HourlyForecastDto.kt
@@ -5,14 +5,13 @@ import com.google.gson.annotations.SerializedName
 data class HourlyForecastDto(
     @SerializedName("time") val time: String? = null,
     @SerializedName("temperature") val temperature: Double? = null,
-    @SerializedName("precipitation_probability") val precipitationProbability: Int? = null,
+    @SerializedName("precipitation_probability") val precipitationProbability: Double? = null,
     @SerializedName("wind_speed") val windSpeed: Double? = null,
     @SerializedName("condition_code") val conditionCode: String? = null,
     @SerializedName("icon") val icon: String? = null,
-    @SerializedName("humidity") val humidity: Int? = null,
+    @SerializedName("humidity") val humidity: Double? = null,
     @SerializedName("feels_like") val feelsLike: Double? = null,
     @SerializedName("wind_gust") val windGust: Double? = null,
-    @SerializedName("uv_index") val uvIndex: Int? = null,
+    @SerializedName("uv_index") val uvIndex: Double? = null,
     @SerializedName("icon_path") val iconPath: String? = null
 )
-
